Add stub so component is always defined.
